[net] Add DNS name resolution to in_gethostbyname and network apps
