+++ elkscmd/inet/urlget/net.c
@@ -18,69 +18,54 @@
 #include <netinet/in.h>
 #include <arpa/inet.h>
 
-void errmsg(const char *, ...);
-
 int net_connect(char *host, int port)
 {
-	int netfd;
+	int netfd, e;
 	struct sockaddr_in in_adr;
 	struct linger l;
-	int ret;
 	
 	netfd = socket(AF_INET, SOCK_STREAM, 0);
-	if (netfd < 0) {
-		errmsg("Can't open socket (check if ktcp is running)");
+	if (netfd < 0)
 		return -1;
-	}
 	
 	in_adr.sin_family = AF_INET;
 	in_adr.sin_port = PORT_ANY;
 	in_adr.sin_addr.s_addr = INADDR_ANY;
 
-    ret = bind(netfd, (struct sockaddr *)&in_adr, sizeof(struct sockaddr_in));
-	if (ret < 0) {
-		perror("Bind failed");
-		close(netfd);
-		return -1;
-	}
+	if (bind(netfd, (struct sockaddr *)&in_adr, sizeof(struct sockaddr_in)) < 0)
+		goto error;
 	
 	l.l_onoff = 1;	/* turn on linger option: will send RST on close*/
 	l.l_linger = 0;	/* must be 0 to turn on option*/
-	ret = setsockopt(netfd, SOL_SOCKET, SO_LINGER, &l, sizeof(l));
-	if (ret < 0)
-		perror("SO_LINGER RST");
+	setsockopt(netfd, SOL_SOCKET, SO_LINGER, &l, sizeof(l));
 
 	in_adr.sin_family = AF_INET;
 	in_adr.sin_port = htons(port);
 	in_adr.sin_addr.s_addr = in_gethostbyname(host);
-	if (in_adr.sin_addr.s_addr == 0) {
-		errmsg("Can't find host '%s' in /etc/hosts", host);
-		close(netfd);
-		return -1;
-	}
+	if (!in_adr.sin_addr.s_addr)
+		goto error;
 
-	ret = connect(netfd, (struct sockaddr *)&in_adr, sizeof(struct sockaddr_in));
-	if (ret < 0){
-		close(netfd);
-		return -1;
-	}
+	if (in_connect(netfd, (struct sockaddr *)&in_adr, sizeof(struct sockaddr_in), 10) < 0)
+		goto error;
 
-	return(netfd);
+	return netfd;
+error:
+	e = errno;
+	close(netfd);
+	errno = e;
+	return -1;
 }
 
 /* if errflag set, send TCP RST on close, else send FIN */
 void net_close(int fd, int errflag)
 {
-	int ret;
-	struct linger l;
-
 	/* Send RST on close was previously turned on by net_connect*/
 	if (!errflag) {
+		struct linger l;
+
 		l.l_onoff = 0;	/* turn off linger option: will send FIN on close*/
 		l.l_linger = 0;
-		ret = setsockopt(fd, SOL_SOCKET, SO_LINGER, &l, sizeof(l));
-		if (ret < 0)
-			perror("SO_LINGER FIN");
+		setsockopt(fd, SOL_SOCKET, SO_LINGER, &l, sizeof(l));
 	}
 	close(fd);
 }
