+++ elkscmd/rootfs_template/bootopts
@@ -1,5 +1,5 @@
 ## boot options max size 511 bytes
-#console=ttyS0,57600 debug net=eth 3 # serial console, multiuser, networking
+console=ttyS0,57600 debug net=eth 3 # serial console, multiuser, networking
 #QEMU=1			# to use ftp/ftpd in qemu
 #TZ=MDT7
 #LOCALIP=10.0.2.16
