+++ libc/include/arpa/inet.h
@@ -2,6 +2,8 @@
 #include <sys/socket.h>
 
 ipaddr_t in_aton(const char *str);
+int	     in_connect(int socket, const struct sockaddr *address, socklen_t address_len,
+			int secs);
 ipaddr_t in_gethostbyname(char *str);
-int in_connect(int socket, const struct sockaddr *address, socklen_t address_len, int secs);
-char *in_ntoa(ipaddr_t in);
+char *   in_ntoa(ipaddr_t in);
+ipaddr_t in_resolv(char *hostname, char *server);
